+++ src/main/java/com/ats/webapi/repository/tally/TallySVoucherUpdateRepository.java
@@ -17,6 +17,6 @@ public interface TallySVoucherUpdateRepository extends JpaRepository<SalesVouche
 	@Transactional
 	@Modifying
 	@Query("UPDATE SalesVouchers s SET s.isTallySync =:isTallySync  WHERE s.billNo=:billNo")
-	int updateSalesVouchers(@Param("billNo")int billNo,@Param("isTallySync") int isTallySync);
+	int updateSalesVouchers(@Param("billNo")int billNo,@Param("isTallySync") long isTallySync);
 
 }
