+++ src/main/java/com/ats/webapi/service/tally/SalesVoucherService.java
@@ -12,7 +12,7 @@ public interface SalesVoucherService {
 
 	SalesVoucherList getAllSalesVoucher();
 
-	ErrorMessage updateSalesVouchers(int billNoList, int isTallySync);
+	ErrorMessage updateSalesVouchers(int billNoList, long isTallySync);
 
 
 }
